+++ doc/release/release_notes.rst
@@ -4,6 +4,30 @@
 Release Notes
 *************
 
+.. release:: Upcoming
+
+    .. change:: fixed
+
+        Updated :class:`wiz.graph.Resolver` and :class:`wiz.graph.Graph` to
+        ensure that packages added during the conflict resolution process are
+        correctly linked to the correct parents instead of
+        :attr:`root <wiz.graph.Graph.ROOT>`.
+
+.. release:: 2.1.0
+    :date: 2019-02-11
+
+    .. change:: changed
+
+        Updated :func:`wiz.definition.query` to add the following rule when
+        guessing the namespace of a package definition: If several namespaces
+        are available, default to the one which is identical to the identifier
+        if possible.
+
+        For instance, the following command will default to ``massive::massive``
+        even if ``maya::massive`` is available::
+
+            >> wiz use massive
+
 .. release:: 2.0.0
     :date: 2019-02-04